refactor immich asset grid to use ref and provider
