+++ mobile/lib/modules/search/providers/search_result_page.provider.dart
@@ -1,10 +1,8 @@
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:immich_mobile/modules/home/ui/asset_grid/asset_grid_data_structure.dart';
+import 'package:immich_mobile/modules/asset_viewer/providers/render_list.provider.dart';
 import 'package:immich_mobile/modules/search/models/search_result_page_state.model.dart';
 
 import 'package:immich_mobile/modules/search/services/search.service.dart';
-import 'package:immich_mobile/modules/settings/providers/app_settings.provider.dart';
-import 'package:immich_mobile/modules/settings/services/app_settings.service.dart';
 import 'package:immich_mobile/shared/models/asset.dart';
 
 class SearchResultPageNotifier extends StateNotifier<SearchResultPageState> {
@@ -55,15 +53,6 @@ final searchResultPageProvider =
 });
 
 final searchRenderListProvider = FutureProvider((ref) {
-  var settings = ref.watch(appSettingsServiceProvider);
-
   final assets = ref.watch(searchResultPageProvider).searchResult;
-
-  final layout = AssetGridLayoutParameters(
-    settings.getSetting(AppSettingsEnum.tilesPerRow),
-    settings.getSetting(AppSettingsEnum.dynamicLayout),
-    GroupAssetsBy.values[settings.getSetting(AppSettingsEnum.groupAssetsBy)],
-  );
-
-  return RenderList.fromAssets(assets, layout);
+  return ref.watch(renderListProvider(assets));
 });
